Fixed bug on TransformParams method and add this and the method ListToVector to ParametrizedModel.py
