+++ Studies/config/params.json
@@ -2,6 +2,6 @@
     "num_den_layers_pre" : 0 , "num_neurons_den_layers_pre": 48, "dropout_rate_den_layers_pre":  0.0734,
     "num_den_layers_post": 1, "num_neurons_den_layers_post": 59, "dropout_rate_den_layers_post": 0.1512,
     "num_lstm_layers": 1, "num_neurons_lstm_layer":100, "activation_dense_pre": 1,
-    "activation_dense_post": 1, "dropout_rate_lstm": 0.04617, "batch_size":1000, "learning_rate":0.1,
+    "activation_dense_post": 1, "dropout_rate_lstm": 0.04617, "batch_size":1000, "learning_rate":0.001,
     "lstm_activation": 1, "lstm_recurrent_activation": 1
 }
